Move fill_in() and helpers to new file fillin.py.
